==> 4_eda.ipynb <==
# Exploratory Data Analysis

## First, the data and tools imports

import pandas as pd
from pprint import pprint
import matplotlib.pyplot as plt
plt.style.use('ggplot')
%matplotlib inline

df = pd.read_table('data/preprocessed.tsv')

# Exercise 1: Go through each column and figure out it's characteristics

pprint(df.columns.tolist())

## Exercise 1.1: "Describe" the columns which have numerical types
### _Hint_: use the `describe` method
### Sample code:

df['n_votes'].describe()

### Try it out on other columns

# ENTER CODE HERE

## Exercise 1.2: Is there a way to compactly visualize summary statistics of a column?
### _Hint_: Use a histogram
### Sample code:

df['n_votes'].hist(bins=20)

### Try it out on other columns

# ENTER CODE HERE

# Exercise 2: Problem - What about columns (features) that are non-numerical?
## _Hint_: Try different kind of plots
## Sample code:

df['section'].value_counts().plot(kind="bar")

df['section'].value_counts().plot(kind="pie")

### Try it out on other (non-numerical) columns

### ENTER CODE HERE

# Exercise 3: What about multivariate analysis? How does one column affect the other?
## Exercise 3.1: Is there a relation between the number of votes and whether the talk is selected?
### _Hint_: Use a scatterplot
### Sample code:

df.plot.scatter(x="n_votes", y="selected")

## Exercise 3.2: Find correlation between other pairs of columns

### ENTER CODE HERE
